% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_indeed.R
\name{in_scrape}
\alias{in_scrape}
\title{Scrape Indeed}
\usage{
in_scrape(q, l, n = 10, base.url = "http://www.indeed.com",
  sleep = sample(20:30, 1), verbose = FALSE)
}
\arguments{
\item{q}{query keyword, "what" on official website. Required}

\item{l}{location, "where" on official website. Optional}

\item{n}{number of results requested}

\item{base.url}{base url for queries defaults to global \code{http://www.indeed.com}}

\item{sleep}{break between queries in seconds, applies when requesting more than the default \code{10} results (\code{n})}

\item{verbose}{sets verbose of \code{rvest::read_html}}
}
\value{
returns data.frame with following variables:
\itemize{
 \item{job title}
 \item{company}
 \item{summary}
 \item{date}
 \item{location}
 \item{url} 
}
}
\description{
Scrapes indeed.com
}
\examples{
\dontrun{
ds <- in_scrape("data scientist", "United States")
da <- in_scrape("data analyst", "New York", 30)
ds_cn <- in_scrape("data scientist", "Beijing", 10, 
                   "http://cn.indeed.com/")
}

}

