% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{in_settings}
\alias{in_settings}
\title{Retrieve settings}
\usage{
in_settings()
}
\value{
Returns data.frame of settings:
\itemize{
 \item base.url
 \item country
}
}
\description{
Rerieve settings set with \code{\link{in_setup}}
}
\examples{
in_setup(country = "us") 
(my_settings <- in_settings())

}

