% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_indeed.R
\name{in_scrape}
\alias{in_scrape}
\title{Scrape Indeed}
\usage{
in_scrape(q, l, p = 1, sleep = sample(10:20, 1))
}
\arguments{
\item{q}{query keyword, "what" on official website. Required}

\item{l}{location, "where" on official website. Optional}

\item{p}{number of pages to scrape}

\item{sleep}{break between queries in seconds, 
applies when requesting more than the default one page of results 
(\code{p = 1}), defaults to \code{sample(10:20, 1)}}
}
\value{
returns data.frame with following variables:
\itemize{
 \item{job title}
 \item{company}
 \item{summary}
 \item{date}
 \item{location}
 \item{url} 
}
}
\description{
Scrapes indeed.com
}
\examples{
\dontrun{
in_setup(country = "us") # setup
ds <- in_scrape(q = "data scientist", l = "Washington", p = 3)
}

}
\seealso{
\code{\link{in_setup}}
}

