% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_indeed.R
\name{in_scrape}
\alias{in_scrape}
\title{Scrape Indeed}
\usage{
in_scrape(q, l, n = 10, base.url = "http://www.indeed.com",
  sleep = sample(20:30, 1), verbose = FALSE)
}
\arguments{
\item{q}{query keyword, "what" on official website. Required}

\item{l}{location, "where" on official website. Optional}

\item{n}{number of results requested}

\item{base.url}{base url for queries defaults to global \code{http://www.indeed.com}}

\item{sleep}{break between queries in seconds, applies when requesting more than the default \code{10} results (\code{n})}

\item{verbose}{if \code{TRUE} prints feedback in console}
}
\value{
returns data.frame of
}
\description{
Scrapes indeed.com
}

